+++ pyrates/frontend/template/operator_graph.py
@@ -124,7 +124,7 @@ class OperatorGraphTemplate(AbstractBaseTemplate):
             # if a value for this particular variation has been passed, overwrite the previous value
             if template.name in value_updates:
                 values_to_update.update(value_updates.pop(template.name))
-            # apply operator tempalte to get OperatorIR and associated default values and label
+            # apply operator template to get OperatorIR and associated default values and label
             operator, op_values, key = template.apply(return_key=True, values=values_to_update)
             operators[key] = operator
             all_values[key] = op_values